#41 Move worklog mapper package
