+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/api/jira/adapter/JiraWorklogClientV3Adapter.java
@@ -2,7 +2,7 @@ package com.senla.worklog.reminder.api.jira.adapter;
 
 import com.senla.worklog.reminder.api.jira.v3.client.JiraWorklogClientV3;
 import com.senla.worklog.reminder.model.Worklog;
-import com.senla.worklog.reminder.model.mapper.WorklogV3Mapper;
+import com.senla.worklog.reminder.api.jira.adapter.mapper.WorklogV3Mapper;
 import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDate;
