+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/api/jira/adapter/JiraWorklogClientV4Adapter.java
@@ -2,7 +2,7 @@ package com.senla.worklog.reminder.api.jira.adapter;
 
 import com.senla.worklog.reminder.api.jira.v4.client.JiraWorklogClientV4;
 import com.senla.worklog.reminder.model.Worklog;
-import com.senla.worklog.reminder.model.mapper.WorklogV4Mapper;
+import com.senla.worklog.reminder.api.jira.adapter.mapper.WorklogV4Mapper;
 import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDate;
