+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/api/jira/adapter/mapper/WorklogV3Mapper.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.model.mapper;
+package com.senla.worklog.reminder.api.jira.adapter.mapper;
 
 import com.senla.worklog.reminder.model.Worklog;
 import com.senla.worklog.reminder.api.jira.v3.model.WorklogV3;
