+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/api/jira/adapter/mapper/WorklogV4MapperImpl.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.model.mapper;
+package com.senla.worklog.reminder.api.jira.adapter.mapper;
 
 import com.senla.worklog.reminder.api.jira.v4.model.IssueV4;
 import com.senla.worklog.reminder.api.jira.v4.model.WorklogV4;
