+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/config/JiraWorklogClientAdapterConfig.java
@@ -3,7 +3,7 @@ package com.senla.worklog.reminder.config;
 import com.senla.worklog.reminder.api.jira.adapter.JiraWorklogClientAdapter;
 import com.senla.worklog.reminder.api.jira.adapter.JiraWorklogClientV4Adapter;
 import com.senla.worklog.reminder.api.jira.v4.client.JiraWorklogClientV4;
-import com.senla.worklog.reminder.model.mapper.WorklogV4Mapper;
+import com.senla.worklog.reminder.api.jira.adapter.mapper.WorklogV4Mapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
